+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.0
 
+* Alter Grub boot configuration with `node.sys.boot`. 
 * System login banner with attributes in `node.sys.banner`,
 * Load Linux kernel modules with `linux_module`.
 * Reboot/Shutdown node with `sys_shutdown`.